requirements.txt
tkinter
opencv-python
sqlite
numpy
Pillow
pyttsx3

